+++ include/NeoFOAM/finiteVolume/cellCentred.hpp
@@ -2,7 +2,7 @@
 // SPDX-FileCopyrightText: 2024 NeoFOAM authors
 
 #include "cellCentred/boundary/boundaryPatchMixin.hpp"
-#include "cellCentred/boundary/surfaceBoundaryBase.hpp"
+#include "cellCentred/boundary/surfaceBoundaryFactory.hpp"
 #include "cellCentred/boundary/volumeBoundaryFactory.hpp"
 
 #include "cellCentred/fields/geometricField.hpp"